Adds Tree
This represents a collection of bags, but... isn't quite a tree. It
behaves like a tree, but there is some work to do to actually get the
many to many relationships represented or at least - as mentioned in the
previous commit - to get a dictionary going to make this more efficient.